pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agentmesh"
version = "0.1.0"
description = "Python client SDK for AgentMesh-Net protocol v0.1"
requires-python = ">=3.10"
dependencies = [
    "jcs>=0.2.1",
    "pynacl>=1.5.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
include = ["agentmesh*"]
